+++ video/common/psb_video_drv.h
@@ -249,7 +249,7 @@ struct psb_validate_buffer {
 struct psb_video_ctx {
 	struct list_head head;
 	struct file *filp; /* DRM device file pointer */
-	int ctx_type; /* (msvdx_tile&0xff)<<16|profile<<8|entrypoint */
+	uint64_t ctx_type; /* (msvdx_tile&0xff)<<16|profile<<8|entrypoint */
 	/* todo: more context specific data for multi-context support */
 	/* Context save and restore */
 	struct ttm_buffer_object *reg_saving_bo;
